Retrieve list of words for a particular domain
To kick things off, we're querying the Oxford Dictionary API to retrieve
a list of words filtered on a particular domain (subject area for a word
such as Music, Computing, Architecture).

For now, we're getting the domain to filter on as a command-line
argument and after querying for a list, print the number of entries
returned.

[Issue ref: garg000dhruv/wordoftheday#1]